Mapping funcionando(ya mete en tablas intermedias y se puede hacer una reserva completa sin tocar ninguna otra tabla que la de book)

+++ src/main/java/com/organization/naturecitytours/trip/TripController.java
@@ -44,7 +44,7 @@ public class TripController {
      * @return a ModelMap with the model attributes for the view
      */
     @RequestMapping("/trip/{tripId}")
-    public ModelAndView showOTrip(@PathVariable("tripId") Long tripId) {
+    public ModelAndView showOTrip(@PathVariable("tripId") int tripId) {
         ModelAndView mav = new ModelAndView("trip/trip");
